Fix js error. Fix aligment for vouchers and reward points.
Fix js error that appears due to adding some file to controller/extension/d_quickcheckout/* . Reason: this file become a 'step' for checkout, and due to processing all steps in the front end, it evokes error that didn't permit saving changes in the editor.
Fixed alignment for voucher and reward points.